requirements.txt
grpcio>=1.17.1,<2.0.0
numpy>=1.15.0,<2.0.0
scikit-image>=0.14.0,<1.0.0
python-decouple>=3.1,<4.0
